+++ examples/live/binance_spot_ema_cross.py
@@ -38,6 +38,9 @@ from nautilus_trader.live.node import TradingNode
 config_node = TradingNodeConfig(
     trader_id="TESTER-001",
     log_level="INFO",
+    exec_engine={
+        "recon_lookback_mins": 1440,
+    },
     # cache_database=CacheDatabaseConfig(),
     data_clients={
         "BINANCE": BinanceDataClientConfig(