+++ tests/integration_tests/adapters/betfair/test_betfair_parsing.py
@@ -204,7 +204,7 @@ class TestBetfairParsing:
         assert result == expected
 
     def test_make_order_limit_on_close(self):
-        order = BetfairTestStubs.limit_order(time_in_force=TimeInForce.ON_CLOSE)
+        order = BetfairTestStubs.limit_order(time_in_force=TimeInForce.AT_THE_CLOSE)
         result = make_order(order)
         expected = {
             "limitOnCloseOrder": {"price": "3.05", "liability": "10.0"},
@@ -246,7 +246,7 @@ class TestBetfairParsing:
     )
     def test_make_order_market_on_close(self, side, liability):
         order = BetfairTestStubs.market_order(
-            time_in_force=TimeInForce.ON_CLOSE, side=OrderSideParser.from_str_py(side)
+            time_in_force=TimeInForce.AT_THE_CLOSE, side=OrderSideParser.from_str_py(side)
         )
         result = make_order(order)
         expected = {