+++ tests/integration_tests/adapters/binance/test_parsing_ws.py
@@ -17,7 +17,7 @@ import pkgutil
 
 import orjson
 
-from nautilus_trader.adapters.binance.parsing.websocket import parse_ticker_24hr_spot_ws
+from nautilus_trader.adapters.binance.parsing.ws_data import parse_ticker_24hr_spot_ws
 from nautilus_trader.backtest.data.providers import TestInstrumentProvider